Fix all psr-4 issues
