+++ src/Util/AuthError.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Supabase\Util;
+
+class AuthError extends Error
+{
+    protected $isAuthError = true;
+
+    public function __construct($message)
+    {
+        super($message);
+        $this->name = 'AuthError';
+    }
+}
